image-decodes/image-decode-path-changes.html

<!DOCTYPE html>
<meta charset=utf-8>
<meta name="timeout" content="long">
<title>HTMLImageElement.decode(), src/srcset mutation tests.</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<link rel=help href="https://html.spec.whatwg.org/multipage/embedded-content.html#dom-img-decode">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

// src tests
// -------------------
promise_test(function(t) {
  var img = new Image();
  img.src = "/images/green.png";
  var promise = img.decode();
  img.src = "/images/green.svg";
  return promise_rejects(t, "EncodingError", promise);
}, document.title + " src changes fail decode.");

promise_test(function(t) {
  var img = new Image();
  img.src = "/images/green.png";
  var first_promise = img.decode();
  img.src = "/images/green.svg";
  var second_promise = img.decode();
  assert_not_equals(first_promise, second_promise);
  return Promise.all([
    promise_rejects(t, "EncodingError", first_promise),
    second_promise
  ]);
}, document.title + " src changes fail decode; following good decode succeeds.");

promise_test(function(t) {
  var img = new Image();
  img.src = "/images/green.png";
  var first_promise = img.decode();
  img.src = "/non/existent/path.png";
  var second_promise = img.decode();
  assert_not_equals(first_promise, second_promise);
  return Promise.all([
    promise_rejects(t, "EncodingError", first_promise),
    promise_rejects(t, "EncodingError", second_promise)
  ]);
}, document.title + " src changes fail decode; following bad decode fails.");

// srcset tests
// -------------------
promise_test(function(t) {
  var img = new Image();
  img.srcset = "/images/green.png 100w";
  var promise = img.decode();
  img.srcset = "/images/green.svg 100w";
  return promise_rejects(t, "EncodingError", promise);
}, document.title + " srcset changes fail decode.");

promise_test(function(t) {
  var img = new Image();
  img.srcset = "/images/green.png 100w";
  var first_promise = img.decode();
  img.srcset = "/images/green.svg 100w";
  var second_promise = img.decode();
  assert_not_equals(first_promise, second_promise);
  return Promise.all([
    promise_rejects(t, "EncodingError", first_promise),
    second_promise
  ]);
}, document.title + " srcset changes fail decode; following good decode succeeds.");

promise_test(function(t) {
  var img = new Image();
  img.srcset = "/images/green.png 100w";
  var first_promise = img.decode();
  img.srcset = "/non/existent/path.png 100w";
  var second_promise = img.decode();
  assert_not_equals(first_promise, second_promise);
  return Promise.all([
    promise_rejects(t, "EncodingError", first_promise),
    promise_rejects(t, "EncodingError", second_promise)
  ]);
}, document.title + " srcset changes fail decode; following bad decode fails.");

</script>
